djangoBackend/requirements.txt
python
django-4.1.6
django_rest_framework
pillow